살려주세요...

+++ app/src/main/java/com/example/cho6/guru2_android12/friendList.kt
@@ -243,7 +243,7 @@ class FriendListViewModel : ViewModel() {
                             val yourLatitude = document.data["latitude"] as Double
                             val yourLongitude = document.data["longitude"] as Double
                             val distance: Double =
-                                getDistance(myLatitude, myLongitude, yourLatitude, yourLongitude)
+                                getDistance(myLatitude, myLongitude, yourLatitude.toDouble(), yourLongitude)
                             if (distance <= 300) {
                                 Log.d("test", "" + document.data["username"])
                                 data.add(
